+++ src/Application/Product/Grid/GridRepository.php
@@ -17,6 +17,10 @@ interface GridRepository
 
     public function filterByProductIds(array $product_ids): static;
 
+    public function filterByVariantIds(array $variant_ids): static;
+
+    public function filterByExplicitVariantIds(array $variant_ids): static;
+
     public function filterByPrice(?string $minimumPriceAmount = null, ?string $maximumPriceAmount = null): static;
 
     public function sortByLabel(): static;